Abstract Input Output modules
